refactor IdentifierKind, adding `.indices()` and reinforce the tests
