+++ src/raw/alg.rs
@@ -20,6 +20,7 @@ use crate::statics::{ARRAY, BITVEC, BOOL, INT, REAL, REGLAN, STRING, SYMBOL_RE};
 use crate::traits::Contains;
 use dashu::float::DBig;
 use dashu::integer::UBig;
+pub use kind::IdentifierKind;
 use num_order::NumHash;
 use serde::{Deserialize, Serialize};
 use std::fmt::{Display, Formatter, Write};
@@ -28,6 +29,8 @@ use std::ops::{Add, Mul, Sub};
 use yaspar::ast::Keyword;
 use yaspar::tokens::SPECIAL_SYMBOLS;
 use yaspar::{binary_to_string, hex_to_string};
+
+mod kind;
 
 /// Represent a literal constant in SMTLib
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
@@ -108,256 +111,6 @@ impl<Str> Index<Str> {
 pub struct Identifier<Str> {
     pub symbol: Str,
     pub indices: Vec<Index<Str>>,
-}
-
-/// Describe pre-defined kinds for identifiers
-#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub enum IdentifierKind<Str> {
-    // Core
-    And,
-    Or,
-    Not,
-    Implies,
-    Xor,
-    Eq,
-    Distinct,
-    Ite,
-
-    // ArrayEx
-    Select,
-    Store,
-
-    // Ints and Reals
-    Add,
-    /// includes unary minus
-    Sub,
-    Mul,
-    /// div; integer division
-    Idiv,
-    /// /; real division
-    Rdiv,
-    Mod,
-    Abs,
-    Le,
-    Lt,
-    Ge,
-    Gt,
-    ToReal,
-    ToInt,
-    IsInt,
-
-    // Strings
-    Char(Vec<u8>, usize),
-    StrConcat,
-    StrLen,
-    StrLt,
-    StrToRe,
-    StrInRe,
-    ReNone,
-    ReAll,
-    ReAllChar,
-    ReConcat,
-    ReUnion,
-    ReInter,
-    ReStar,
-    StrLe,
-    StrAt,
-    StrSubstr,
-    StrPrefixof,
-    StrSuffixof,
-    StrContains,
-    StrIndexof,
-    StrReplace,
-    StrReplaceAll,
-    StrReplaceRe,
-    StrReplaceReAll,
-    ReComp,
-    ReDiff,
-    ReAdd,
-    ReOpt,
-    ReRange,
-    RePower(UBig),
-    ReLoop(UBig, UBig),
-    StrIsDigit,
-    StrToCode,
-    StrFromCode,
-    StrToInt,
-    StrFromInt,
-
-    // Bit Vectors
-    Concat,
-    Extract(UBig, UBig),
-    BvNot,
-    BvNeg,
-    BvAnd,
-    BvOr,
-    BvAdd,
-    BvMul,
-    BvUdiv,
-    BvUrem,
-    BvShl,
-    BvLshr,
-    BvUlt,
-    BvNego,
-    BvUaddo,
-    BvSaddo,
-    BvUmulo,
-    BvSmulo,
-    /// standard
-    UbvToInt,
-    /// standard
-    SbvToInt,
-    Bv2Nat,
-    Bv2Int,
-    IntToBv(UBig),
-    Nat2Bv(UBig),
-    Int2Bv(UBig),
-    BvNand,
-    BvNor,
-    BvXor,
-    BvNxor,
-    BvComp,
-    BvSub,
-    BvSdiv,
-    BvSrem,
-    BvSmod,
-    BvAShr,
-    BvUsubo,
-    BvSsubo,
-    BvSdivo,
-    Repeat(UBig),
-    ZeroExtend(UBig),
-    SignExtend(UBig),
-    RotateLeft(UBig),
-    RotateRight(UBig),
-    BvUle,
-    BvUgt,
-    BvUge,
-    BvSlt,
-    BvSle,
-    BvSgt,
-    BvSge,
-
-    // datatypes
-    Is(Str),
-}
-
-impl<Str> IdentifierKind<Str> {
-    pub fn name(&self) -> &'static str {
-        match self {
-            IdentifierKind::And => "and",
-            IdentifierKind::Or => "or",
-            IdentifierKind::Not => "not",
-            IdentifierKind::Implies => "=>",
-            IdentifierKind::Xor => "xor",
-            IdentifierKind::Eq => "=",
-            IdentifierKind::Distinct => "distinct",
-            IdentifierKind::Ite => "ite",
-            IdentifierKind::Select => "select",
-            IdentifierKind::Store => "store",
-            IdentifierKind::Add => "+",
-            IdentifierKind::Sub => "-",
-            IdentifierKind::Mul => "*",
-            IdentifierKind::Idiv => "div",
-            IdentifierKind::Rdiv => "/",
-            IdentifierKind::Mod => "mod",
-            IdentifierKind::Abs => "abs",
-            IdentifierKind::Le => "<=",
-            IdentifierKind::Lt => "<",
-            IdentifierKind::Ge => ">=",
-            IdentifierKind::Gt => ">",
-            IdentifierKind::ToReal => "to_real",
-            IdentifierKind::ToInt => "to_int",
-            IdentifierKind::IsInt => "is_int",
-            IdentifierKind::Char(_, _) => "char",
-            IdentifierKind::StrConcat => "str.++",
-            IdentifierKind::StrLen => "str.len",
-            IdentifierKind::StrLt => "str.<",
-            IdentifierKind::StrToRe => "str.to_re",
-            IdentifierKind::StrInRe => "str.in_re",
-            IdentifierKind::ReNone => "re.none",
-            IdentifierKind::ReAll => "re.all",
-            IdentifierKind::ReAllChar => "re.allchar",
-            IdentifierKind::ReConcat => "re.++",
-            IdentifierKind::ReUnion => "re.union",
-            IdentifierKind::ReInter => "re.inter",
-            IdentifierKind::ReStar => "re.*",
-            IdentifierKind::StrLe => "str.<=",
-            IdentifierKind::StrAt => "str.at",
-            IdentifierKind::StrSubstr => "str.substr",
-            IdentifierKind::StrPrefixof => "str.prefixof",
-            IdentifierKind::StrSuffixof => "str.suffixof",
-            IdentifierKind::StrContains => "str.contains",
-            IdentifierKind::StrIndexof => "str.indexof",
-            IdentifierKind::StrReplace => "str.replace",
-            IdentifierKind::StrReplaceAll => "str.replace_all",
-            IdentifierKind::StrReplaceRe => "str.replace_re",
-            IdentifierKind::StrReplaceReAll => "str.replace_re_all",
-            IdentifierKind::ReComp => "re.comp",
-            IdentifierKind::ReDiff => "re.diff",
-            IdentifierKind::ReAdd => "re.+",
-            IdentifierKind::ReOpt => "re.opt",
-            IdentifierKind::ReRange => "re.range",
-            IdentifierKind::RePower(_) => "re.^",
-            IdentifierKind::ReLoop(_, _) => "re.loop",
-            IdentifierKind::StrIsDigit => "str.is_digit",
-            IdentifierKind::StrToCode => "str.to_code",
-            IdentifierKind::StrFromCode => "str.from_code",
-            IdentifierKind::StrToInt => "str.to_int",
-            IdentifierKind::StrFromInt => "str.from_int",
-            IdentifierKind::Concat => "concat",
-            IdentifierKind::Extract(_, _) => "extract",
-            IdentifierKind::BvNot => "bvnot",
-            IdentifierKind::BvNeg => "bvneg",
-            IdentifierKind::BvAnd => "bvand",
-            IdentifierKind::BvOr => "bvor",
-            IdentifierKind::BvAdd => "bvadd",
-            IdentifierKind::BvMul => "bvmul",
-            IdentifierKind::BvUdiv => "bvudiv",
-            IdentifierKind::BvUrem => "bvurem",
-            IdentifierKind::BvShl => "bvshl",
-            IdentifierKind::BvLshr => "bvlshr",
-            IdentifierKind::BvUlt => "bvult",
-            IdentifierKind::BvNego => "bvnego",
-            IdentifierKind::BvUaddo => "bvuaddo",
-            IdentifierKind::BvSaddo => "bvsaddo",
-            IdentifierKind::BvUmulo => "bvumulo",
-            IdentifierKind::BvSmulo => "bvsmulo",
-            IdentifierKind::UbvToInt => "ubv_to_int",
-            IdentifierKind::SbvToInt => "sbv_to_int",
-            IdentifierKind::Bv2Nat => "bv2nat",
-            IdentifierKind::Bv2Int => "bv2int",
-            IdentifierKind::IntToBv(_) => "int_to_bv",
-            IdentifierKind::Nat2Bv(_) => "nat2bv",
-            IdentifierKind::Int2Bv(_) => "int2bv",
-            IdentifierKind::BvNand => "bvnand",
-            IdentifierKind::BvNor => "bvnor",
-            IdentifierKind::BvXor => "bvxor",
-            IdentifierKind::BvNxor => "bvnxor",
-            IdentifierKind::BvComp => "bvcomp",
-            IdentifierKind::BvSub => "bvsub",
-            IdentifierKind::BvSdiv => "bvsdiv",
-            IdentifierKind::BvSrem => "bvsrem",
-            IdentifierKind::BvSmod => "bvsmod",
-            IdentifierKind::BvAShr => "bvashr",
-            IdentifierKind::BvUsubo => "bvusubo",
-            IdentifierKind::BvSsubo => "bvssubo",
-            IdentifierKind::BvSdivo => "bvsdivo",
-            IdentifierKind::Repeat(_) => "repeat",
-            IdentifierKind::ZeroExtend(_) => "zero_extend",
-            IdentifierKind::SignExtend(_) => "sign_extend",
-            IdentifierKind::RotateLeft(_) => "rotate_left",
-            IdentifierKind::RotateRight(_) => "rotate_right",
-            IdentifierKind::BvUle => "bvule",
-            IdentifierKind::BvUgt => "bvugt",
-            IdentifierKind::BvUge => "bvuge",
-            IdentifierKind::BvSlt => "bvslt",
-            IdentifierKind::BvSle => "bvsle",
-            IdentifierKind::BvSgt => "bvsgt",
-            IdentifierKind::BvSge => "bvsge",
-            IdentifierKind::Is(_) => "is",
-        }
-    }
 }
 
 /// This trait checks whether som identifier is a given string
